requirements.txt
fastapi==0.103.1
pydantic==2.3.0
uvicorn[standard]==0.23.0
sqlalchemy==2.0.20
asyncpg==0.28.0
alembic==1.12.0
pandas==2.1.1
python-dotenv==1.0.0
loguru==0.7.0
greenlet==3.1.1
pytest
pytest-asyncio
